+++ data/src/message.rs
@@ -13,7 +13,10 @@ use serde::{Deserialize, Deserializer, Serialize};
 use url::Url;
 
 pub use self::formatting::Formatting;
-pub use self::source::Source;
+pub use self::source::{
+    server::{Kind, StandardReply},
+    Source,
+};
 
 use crate::config::buffer::UsernameFormat;
 use crate::time::{self, Posix};
@@ -178,8 +181,7 @@ impl Message {
                 Source::Server(Some(server)) => {
                     matches!(
                         server.kind(),
-                        source::server::Kind::MonitoredOnline
-                            | source::server::Kind::MonitoredOffline
+                        Kind::MonitoredOnline | Kind::MonitoredOffline | Kind::StandardReply(_)
                     )
                 }
                 Source::Internal(source::Internal::Logs) => true,
@@ -195,10 +197,10 @@ impl Message {
         } else if let Source::Server(Some(source)) = self.target.source() {
             if matches!(
                 source.kind(),
-                source::server::Kind::ReplyTopic
-                    | source::server::Kind::MonitoredOnline
-                    | source::server::Kind::MonitoredOffline
-                    | source::server::Kind::ChangeHost
+                Kind::ReplyTopic
+                    | Kind::MonitoredOnline
+                    | Kind::MonitoredOffline
+                    | Kind::ChangeHost
             ) {
                 return false;
             }
@@ -694,7 +696,7 @@ fn target(
 
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(None),
+                source: Source::Server(None),
             })
         }
         Command::TOPIC(channel, _) | Command::KICK(channel, _, _) => {
@@ -703,7 +705,7 @@ fn target(
 
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(None),
+                source: Source::Server(None),
             })
         }
         Command::PART(channel, _) => {
@@ -712,8 +714,8 @@ fn target(
 
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(Some(source::Server::new(
-                    source::server::Kind::Part,
+                source: Source::Server(Some(source::Server::new(
+                    Kind::Part,
                     Some(user?.nickname().to_owned()),
                 ))),
             })
@@ -724,8 +726,8 @@ fn target(
 
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(Some(source::Server::new(
-                    source::server::Kind::Join,
+                source: Source::Server(Some(source::Server::new(
+                    Kind::Join,
                     Some(user?.nickname().to_owned()),
                 ))),
             })
@@ -735,10 +737,7 @@ fn target(
                 target::Channel::parse(params.get(1)?, chantypes, statusmsg, casemapping).ok()?;
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(Some(source::Server::new(
-                    source::server::Kind::ReplyTopic,
-                    None,
-                ))),
+                source: Source::Server(Some(source::Server::new(Kind::ReplyTopic, None))),
             })
         }
         Command::Numeric(RPL_CHANNELMODEIS, params) => {
@@ -746,7 +745,7 @@ fn target(
                 target::Channel::parse(params.get(1)?, chantypes, statusmsg, casemapping).ok()?;
             Some(Target::Channel {
                 channel,
-                source: source::Source::Server(None),
+                source: Source::Server(None),
             })
         }
         Command::Numeric(RPL_AWAY, params) => {
@@ -833,20 +832,32 @@ fn target(
             }
         }
         Command::CHGHOST(_, _) => Some(Target::Server {
-            source: source::Source::Server(Some(source::Server::new(
-                source::server::Kind::ChangeHost,
+            source: Source::Server(Some(source::Server::new(
+                Kind::ChangeHost,
                 user.map(|user| user.nickname().to_owned()),
             ))),
         }),
         Command::Numeric(RPL_MONONLINE, _) => Some(Target::Server {
-            source: source::Source::Server(Some(source::Server::new(
-                source::server::Kind::MonitoredOnline,
+            source: Source::Server(Some(source::Server::new(Kind::MonitoredOnline, None))),
+        }),
+        Command::Numeric(RPL_MONOFFLINE, _) => Some(Target::Server {
+            source: Source::Server(Some(source::Server::new(Kind::MonitoredOffline, None))),
+        }),
+        Command::FAIL(_, _, _, _) => Some(Target::Server {
+            source: Source::Server(Some(source::Server::new(
+                Kind::StandardReply(StandardReply::Fail),
                 None,
             ))),
         }),
-        Command::Numeric(RPL_MONOFFLINE, _) => Some(Target::Server {
-            source: source::Source::Server(Some(source::Server::new(
-                source::server::Kind::MonitoredOffline,
+        Command::WARN(_, _, _, _) => Some(Target::Server {
+            source: Source::Server(Some(source::Server::new(
+                Kind::StandardReply(StandardReply::Warn),
+                None,
+            ))),
+        }),
+        Command::NOTE(_, _, _, _) => Some(Target::Server {
+            source: Source::Server(Some(source::Server::new(
+                Kind::StandardReply(StandardReply::Note),
                 None,
             ))),
         }),
@@ -1211,6 +1222,36 @@ fn content<'a>(
                 Some(plain(format!("Chat history for {target} at {timestamp}")))
             } else {
                 None
+            }
+        }
+        Command::FAIL(command, _, context, description) => {
+            if let Some(context) = context {
+                Some(plain(format!(
+                    "{command} ({}) failed: {description}",
+                    context.join(", ")
+                )))
+            } else {
+                Some(plain(format!("{command} failed: {description}")))
+            }
+        }
+        Command::WARN(command, _, context, description) => {
+            if let Some(context) = context {
+                Some(plain(format!(
+                    "{command} ({}) warning: {description}",
+                    context.join(", ")
+                )))
+            } else {
+                Some(plain(format!("{command} warning: {description}")))
+            }
+        }
+        Command::NOTE(command, _, context, description) => {
+            if let Some(context) = context {
+                Some(plain(format!(
+                    "{command} ({}) notice: {description}",
+                    context.join(", ")
+                )))
+            } else {
+                Some(plain(format!("{command} notice: {description}")))
             }
         }
         Command::Numeric(_, responses) | Command::Unknown(_, responses) => Some(parse_fragments(