+++ irc/proto/src/command.rs
@@ -98,6 +98,7 @@ pub enum Command {
     /// <accountname>
     ACCOUNT(String),
     /* IRC extensions */
+    /// <type> [<params>]
     BATCH(String, Vec<String>),
     /// <subcommand> [<params>]
     CHATHISTORY(String, Vec<String>),
@@ -117,6 +118,14 @@ pub enum Command {
     TAGMSG(String),
     /// <nickname>
     USERIP(String),
+
+    /* Standard Replies */
+    /// <command> <code> [<context>] <description>
+    FAIL(String, String, Option<Vec<String>>, String),
+    /// <command> <code> [<context>] <description>
+    WARN(String, String, Option<Vec<String>>, String),
+    /// <command> <code> [<context>] <description>
+    NOTE(String, String, Option<Vec<String>>, String),
 
     Numeric(Numeric, Vec<String>),
     Unknown(String, Vec<String>),
@@ -214,6 +223,39 @@ impl Command {
             "MONITOR" if len > 0 => MONITOR(req!(), opt!()),
             "TAGMSG" if len > 0 => TAGMSG(req!()),
             "USERIP" if len > 0 => USERIP(req!()),
+            "FAIL" if len > 2 => {
+                let a = req!();
+                let b = req!();
+                let mut c: Vec<String> = params.collect();
+                let d = c.pop().unwrap();
+                if !c.is_empty() {
+                    FAIL(a, b, Some(c), d)
+                } else {
+                    FAIL(a, b, None, d)
+                }
+            }
+            "WARN" if len > 2 => {
+                let a = req!();
+                let b = req!();
+                let mut c: Vec<String> = params.collect();
+                let d = c.pop().unwrap();
+                if !c.is_empty() {
+                    WARN(a, b, Some(c), d)
+                } else {
+                    WARN(a, b, None, d)
+                }
+            }
+            "NOTE" if len > 2 => {
+                let a = req!();
+                let b = req!();
+                let mut c: Vec<String> = params.collect();
+                let d = c.pop().unwrap();
+                if !c.is_empty() {
+                    NOTE(a, b, Some(c), d)
+                } else {
+                    NOTE(a, b, None, d)
+                }
+            }
             _ => Self::Unknown(tag, params.collect()),
         }
     }
@@ -276,6 +318,21 @@ impl Command {
             Command::MONITOR(a, b) => std::iter::once(a).chain(b).collect(),
             Command::TAGMSG(a) => vec![a],
             Command::USERIP(a) => vec![a],
+            Command::FAIL(a, b, c, d) => std::iter::once(a)
+                .chain(Some(b))
+                .chain(c.into_iter().flatten())
+                .chain(Some(d))
+                .collect(),
+            Command::WARN(a, b, c, d) => std::iter::once(a)
+                .chain(Some(b))
+                .chain(c.into_iter().flatten())
+                .chain(Some(d))
+                .collect(),
+            Command::NOTE(a, b, c, d) => std::iter::once(a)
+                .chain(Some(b))
+                .chain(c.into_iter().flatten())
+                .chain(Some(d))
+                .collect(),
             Command::Numeric(_, params) => params,
             Command::Unknown(_, params) => params,
             Command::Raw(_) => vec![],
@@ -337,6 +394,9 @@ impl Command {
             MONITOR(_, _) => "MONITOR".to_string(),
             TAGMSG(_) => "TAGMSG".to_string(),
             USERIP(_) => "USERIP".to_string(),
+            FAIL(_, _, _, _) => "FAIL".to_string(),
+            WARN(_, _, _, _) => "WARN".to_string(),
+            NOTE(_, _, _, _) => "NOTE".to_string(),
             Numeric(numeric, _) => format!("{:03}", *numeric as u16),
             Unknown(tag, _) => tag.clone(),
             Raw(_) => "".to_string(),